% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aceEditor}
\alias{aceEditor}
\title{Render Ace}
\usage{
aceEditor(outputId, value, mode, theme, vimKeyBinding = FALSE,
  readOnly = FALSE, height = "400px", fontSize = 12, debounce = 1000,
  wordWrap = FALSE, showLineNumbers = TRUE, highlightActiveLine = TRUE,
  selectionId = NULL, cursorId = NULL, hotkeys = NULL)
}
\arguments{
  \item{outputId}{The ID associated with this element}

  \item{value}{The initial text to be contained in the
  editor.}

  \item{mode}{The Ace \code{mode} to be used by the editor.
  The \code{mode} in Ace is often the programming or markup
  language that you're using and determines things like
  syntax highlighting and code folding. Use the
  \code{\link{getAceModes}} function to enumerate all the
  modes available.}

  \item{theme}{The Ace \code{theme} to be used by the
  editor. The \code{theme} in Ace determines the styling
  and coloring of the editor. Use
  \code{\link{getAceThemes}} to enumerate all the themes
  available.}

  \item{readOnly}{If set to \code{TRUE}, Ace will disable
  client-side editing. If \code{FALSE} (the default), it
  will enable editing.}

  \item{height}{A number (which will be interpreted as a
  number of pixels) or any valid CSS dimension (such as
  "\code{50\%}", "\code{200px}", or "\code{auto}").}

  \item{fontSize}{Defines the font size (in px) used in the
  editor and should be an integer. The default is 12.}

  \item{debounce}{The number of milliseconds to debounce
  the input. This will cause the client to withhold update
  notifications until the user has stopped typing for this
  amount of time. If 0, the server will be notified of
  every keystroke as it happens.}

  \item{wordWrap}{If set to \code{TRUE}, Ace will enable
  word wrapping. Default value is \code{FALSE}.}

  \item{cursorId}{The ID associated with a cursor change.}

  \item{selectionId}{The ID associated with a change of
  selected text}

  \item{hotkeys}{A list whose names are ID names and whose
  elements are the shortcuts of keys. Shortcuts can either
  be a simple string or a list with elements 'win' and
  'mac' that that specifies different shortcuts for win and
  mac (see example).}
}
\description{
Render an Ace editor on an application page.
}
\examples{
\dontrun{
 aceEditor("myEditor", "Initial text for editor here", mode="r",
   theme="ambiance")

 aceEditor("myCodeEditor", "# Enter code", mode="r",
   hotkeys = list(helpKey="F1",
                runKey=list(win="Ctrl-R|Ctrl-Shift-Enter",
                            mac="CMD-ENTER|CMD-SHIFT-ENTER")
                ),
   wordWrap=TRUE, debounce=10)
}
}
\author{
Jeff Allen \email{jeff@trestletech.com}
}

