% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-tooltips.R
\name{build_tooltip_fields}
\alias{build_tooltip_fields}
\title{Build the fields used to make an html tooltip}
\usage{
build_tooltip_fields(v)
}
\arguments{
\item{v}{Autocomplete metadata values used for building tooltip info}
}
\value{
a list with html-formatted character values "title" and "body
}
\description{
Build the fields used to make an html tooltip
}
