% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.fname_regex}
\alias{.fname_regex}
\title{Regular expression for matching the function name in a completion line in the
middle of a function call}
\format{
An object of class \code{character} of length 1.
}
\usage{
.fname_regex
}
\description{
Regular expression for matching the function name in a completion line in the
middle of a function call
}
\keyword{datasets}
