% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.tools}
\alias{.tools}
\title{Get namespace to get access to unexported functions, namely
  RdTags}
\format{An object of class \code{environment} of length 727.}
\usage{
.tools
}
\description{
Get namespace to get access to unexported functions, namely
  RdTags
}
\keyword{datasets}
