% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.utils}
\alias{.utils}
\title{Get namespace to get access to unexported functions, namely
  .getHelpFile
  .assignLinebuffer
  .assignEnd
  .guessTokenFromLine
  .completeToken}
\format{An object of class \code{environment} of length 521.}
\usage{
.utils
}
\description{
Get namespace to get access to unexported functions, namely
  .getHelpFile
  .assignLinebuffer
  .assignEnd
  .guessTokenFromLine
  .completeToken
}
\keyword{datasets}
