% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_desc_help}
\alias{get_desc_help}
\title{Retrieve description section from help document}
\usage{
get_desc_help(...)
}
\arguments{
\item{...}{Arguments passed on to \code{get_help_file}
\describe{
  \item{}{}
}}
}
\value{
a character value representing the description section of a help
  document, rendered as HTML
}
\description{
Retrieve description section from help document
}
\examples{
shinyAce:::get_desc_help("match", package = "base")

}
