% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ace-themes.R
\name{getAceThemes}
\alias{getAceThemes}
\title{Get available themes}
\usage{
getAceThemes()
}
\description{
Gets all of the available \code{themes} available in the installed version
of shinyAce. Themes determine the styling and colors used in the editor.
}
\author{
Jeff Allen \email{jeff@trestletech.com}
}
