% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{is.empty}
\alias{is.empty}
\title{Check if vector is empty}
\usage{
is.empty(x)
}
\arguments{
\item{x}{vector}
}
\description{
Check if vector is empty
}
\examples{
is.empty(NULL)
is.empty(NA)
is.empty(c())
is.empty("")
is.empty(" ")
is.empty(c(" ", " "))
is.empty(list())
is.empty(list(a = "", b = ""))
}
