% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js-quote.R
\name{jsQuote}
\alias{jsQuote}
\title{Escape a JS String}
\usage{
jsQuote(text)
}
\arguments{
\item{text}{The text to escape}
}
\description{
Escape a String to be sent to JavaScript
}
\author{
Jeff Allen \email{jeff@trestletech.com}
}

