% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-autocomplete.R
\name{r_completions_function_call_metadata}
\alias{r_completions_function_call_metadata}
\title{R completions when cursor is within a function call}
\usage{
r_completions_function_call_metadata(fname, completions)
}
\arguments{
\item{fname}{the function name for which the function call specific
completion metadata should be constructed}

\item{completions}{a character vector of completions. These will serve as the
foundation for building added R-specific metadata}
}
\description{
R completions when cursor is within a function call
}
