% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-autocomplete.R
\name{r_completions_general_metadata}
\alias{r_completions_general_metadata}
\title{R completions for general case}
\usage{
r_completions_general_metadata(completions)
}
\arguments{
\item{completions}{a character vector of completions. These will serve as the
foundation for building added R-specific metadata}
}
\description{
R completions for general case
}
