% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-autocomplete.R
\name{r_completions_metadata}
\alias{r_completions_metadata}
\title{Return completions for a given line of text}
\usage{
r_completions_metadata(line)
}
\arguments{
\item{line}{the text up until the cursor in the line for autocompletion}
}
\description{
Return completions for a given line of text
}
