% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-autocomplete.R
\name{r_function_completions_metadata}
\alias{r_function_completions_metadata}
\title{R completions when cursor is within a function call}
\usage{
r_function_completions_metadata(fname, completions)
}
\description{
R completions when cursor is within a function call
}
