% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rd_2_html}
\alias{rd_2_html}
\title{Convert an Rd object to HTML}
\usage{
rd_2_html(...)
}
\arguments{
\item{...}{ additional parameters to pass to \code{\link{parse_Rd}} when
    \code{Rd} is a filename. }
}
\value{
a character value of Rd content rendered as HTML
}
\description{
Convert an Rd object to HTML
}
