% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinyAce-options}
\alias{shinyAce-options}
\title{Options available for shinyAce}
\description{
\describe{
  \item{shinyAce.debug}{
    Logical value to enable or disable debugging messages
    being printed to console. default behavior \code{FALSE}.
  }
}
}
