% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinyAce_debug}
\alias{shinyAce_debug}
\title{Function for handling optional debugging messages}
\usage{
shinyAce_debug(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or (for
    \code{message} only) a single condition object.}
}
\description{
Function for handling optional debugging messages
}
