% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-tooltips.R
\name{tooltip_html}
\alias{tooltip_html}
\title{A helper for formatting a tooltip entry}
\usage{
tooltip_html(title = "", body = "")
}
\arguments{
\item{title}{a character value to use as the title}

\item{body}{an html block to embed as the body of the tooltip}
}
\description{
A helper for formatting a tooltip entry
}
