\name{aceEditor}
\alias{aceEditor}
\title{Render Ace}
\usage{
  aceEditor(outputId, value, mode, theme, readOnly,
    height = "400px")
}
\arguments{
  \item{outputId}{The ID associated with this element}
}
\description{
  Render an Ace editor on an application page.
}
\author{
  Jeff Allen \email{jeff@trestletech.com}
}

