% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_arg_help}
\alias{get_arg_help}
\title{Retrieve argument documentation from help document}
\usage{
get_arg_help(..., args = character())
}
\arguments{
\item{...}{Arguments passed on to \code{get_help_file}
\describe{
  \item{}{}
}}

\item{args}{function arguments names to get documentation for}
}
\value{
A character vector of help
}
\description{
Retrieve argument documentation from help document
}
\examples{
shinyAce:::get_arg_help("match", package = "base", args = c("table", "nomatch"))

}
