% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_help_file}
\alias{get_help_file}
\title{Retrieve an Rd object of a help query}
\usage{
get_help_file(topic, package = NULL, ...)
}
\arguments{
\item{topic}{character name of help topic}

\item{package}{character name of help topic package namespace}

\item{...}{additional arguments passed to \code{utils:::.getHelpFile}}
}
\value{
the Rd object returned from \code{utils:::.getHelpFile}
}
\description{
Safely return NULL if an error is encountered.
}
