% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_help_file}
\alias{get_help_file}
\title{Retrieve an Rd object of a help query}
\usage{
get_help_file(...)
}
\arguments{
\item{...}{arguments passed to \code{utils::help}}
}
\value{
the Rd object returned from utils:::getHelpFile
}
\description{
Safely return NULL if an error is encountered.
}
