% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_usage_help}
\alias{get_usage_help}
\title{Retrieve usage section from help document}
\usage{
get_usage_help(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_help_file]{get_help_file}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
a character value representing the usage section of a help
  document, rendered as HTML
}
\description{
Retrieve usage section from help document
}
\examples{
shinyAce:::get_usage_help("match", package = "base")

}
