% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{re_capture}
\alias{re_capture}
\title{Retrieve regular expression named capture groups as a list}
\usage{
re_capture(x, re, ...)
}
\arguments{
\item{x}{a character string to capture from}

\item{re}{the regular expression to use}

\item{...}{additional arguments passed to \code{\link[base]{regexpr}}}
}
\value{
a named list of matches
}
\description{
Retrieve regular expression named capture groups as a list
}
\examples{
shinyAce:::re_capture("ak09j b", "(?<num>\\\\d+)(?<alpha>[a-zA-Z]+)", perl = TRUE)

}
