\name{updateAceEditor}
\alias{updateAceEditor}
\title{Update Ace Editor}
\usage{
  updateAceEditor(session, editorId, value, theme, mode)
}
\arguments{
  \item{session}{The Shiny session to whom the editor
  belongs}

  \item{editorId}{The ID associated with this element}

  \item{theme}{The Ace theme}

  \item{mode}{The Ace mode}
}
\description{
  Update the styling or mode of an aceEditor component.
}
\author{
  Jeff Allen \email{jeff@trestletech.com}
}

